Move book detail page component (#178)
