+++ frontend/app/routes/home.books.$bookId._index/components/BookDetailBorrower.tsx
