+++ frontend/app/routes/home.books.$bookId._index/components/BookDetailContent.tsx
