+++ frontend/app/routes/home.books.$bookId._index/components/BookDetailTitle.tsx
