+++ frontend/app/routes/home.books.$bookId/components/BookDetailControlButtons.tsx
