+++ frontend/app/routes/home.books.$bookId/components/BookDetailDeleteButton.tsx
