+++ frontend/app/routes/home.books.$bookId/components/BookDetailDeleteModal.tsx
