+++ frontend/app/routes/home.books.$bookId/components/BookDetailEditButton.tsx
