+++ frontend/app/routes/home.books.$bookId/components/BookDetailThumbnail.tsx
